Rename src/deck → src/decks for multi-deck namespace
