+++ src/views/DeckView.test.tsx
@@ -4,8 +4,8 @@ import userEvent from "@testing-library/user-event";
 import type { ReactNode } from "react";
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test } from "vitest";
 import { itemCardFactory } from "../cards/factories";
-import { deckFactory } from "../deck/factories";
-import { useDeckStore } from "../deck/store";
+import { deckFactory } from "../decks/factories";
+import { useDeckStore } from "../decks/store";
 import { magicItemIndexHandler, server } from "../test/msw";
 import { renderWithRouter } from "../test/renderWithRouter";
 import { DeckView } from "./DeckView";
